Convert ssl3_get_record to tls_read_record
We move the old ssl3_get_record function to conform with the new record
layer design.
