+++ ssl/record/ssl3_record.c
@@ -63,45 +63,6 @@ void SSL3_RECORD_set_seq_num(SSL3_RECORD *r, const unsigned char *seq_num)
     memcpy(r->seq_num, seq_num, SEQ_NUM_SIZE);
 }
 
-/*
- * Peeks ahead into "read_ahead" data to see if we have a whole record waiting
- * for us in the buffer.
- */
-static int ssl3_record_app_data_waiting(SSL_CONNECTION *s)
-{
-    SSL3_BUFFER *rbuf;
-    size_t left, len;
-    unsigned char *p;
-
-    rbuf = s->rrlmethod->get0_rbuf(s->rrl);
-
-    p = SSL3_BUFFER_get_buf(rbuf);
-    if (p == NULL)
-        return 0;
-
-    left = SSL3_BUFFER_get_left(rbuf);
-
-    if (left < SSL3_RT_HEADER_LENGTH)
-        return 0;
-
-    p += SSL3_BUFFER_get_offset(rbuf);
-
-    /*
-     * We only check the type and record length, we will sanity check version
-     * etc later
-     */
-    if (*p != SSL3_RT_APPLICATION_DATA)
-        return 0;
-
-    p += 3;
-    n2s(p, len);
-
-    if (left < SSL3_RT_HEADER_LENGTH + len)
-        return 0;
-
-    return 1;
-}
-
 int ossl_early_data_count_ok(SSL_CONNECTION *s, size_t length,
                              size_t overhead, int send)
 {
@@ -149,710 +110,6 @@ int ossl_early_data_count_ok(SSL_CONNECTION *s, size_t length,
     return 1;
 }
 
-static int tls_handle_rlayer_return(SSL_CONNECTION *s, int ret, char *file,
-                                    int line)
-{
-    SSL *ssl = SSL_CONNECTION_GET_SSL(s);
-
-    if (ret == OSSL_RECORD_RETURN_RETRY) {
-        s->rwstate = SSL_READING;
-        ret = -1;
-    } else {
-        s->rwstate = SSL_NOTHING;
-        if (ret == OSSL_RECORD_RETURN_EOF) {
-            if (s->options & SSL_OP_IGNORE_UNEXPECTED_EOF) {
-                SSL_set_shutdown(ssl, SSL_RECEIVED_SHUTDOWN);
-                s->s3.warn_alert = SSL_AD_CLOSE_NOTIFY;
-            } else {
-                ERR_new();
-                ERR_set_debug(file, line, 0);
-                ossl_statem_fatal(s, SSL_AD_DECODE_ERROR,
-                                  SSL_R_UNEXPECTED_EOF_WHILE_READING, NULL);
-            }
-        } else if (ret == OSSL_RECORD_RETURN_FATAL) {
-            ERR_new();
-            ERR_set_debug(file, line, 0);
-            ossl_statem_fatal(s, s->rrlmethod->get_alert_code(s->rrl),
-                              SSL_R_RECORD_LAYER_FAILURE, NULL);
-        }
-        /*
-         * The record layer distinguishes the cases of EOF, non-fatal
-         * err and retry. Upper layers do not.
-         * If we got a retry or success then *ret is already correct,
-         * otherwise we need to convert the return value.
-         */
-        /*
-         * TODO(RECLAYER): What does a non fatal err that isn't a retry even
-         * mean???
-         */
-        if (ret == OSSL_RECORD_RETURN_NON_FATAL_ERR || ret == OSSL_RECORD_RETURN_EOF)
-            ret = 0;
-        else if (ret < OSSL_RECORD_RETURN_NON_FATAL_ERR)
-            ret = -1;
-    }
-
-    return ret;
-}
-
-# define HANDLE_RLAYER_RETURN(s, ret) \
-    tls_handle_rlayer_return(s, ret, OPENSSL_FILE, OPENSSL_LINE)
-
-/*
- * MAX_EMPTY_RECORDS defines the number of consecutive, empty records that
- * will be processed per call to ssl3_get_record. Without this limit an
- * attacker could send empty records at a faster rate than we can process and
- * cause ssl3_get_record to loop forever.
- */
-#define MAX_EMPTY_RECORDS 32
-
-#define SSL2_RT_HEADER_LENGTH   2
-/*-
- * Call this to get new input records.
- * It will return <= 0 if more data is needed, normally due to an error
- * or non-blocking IO.
- * When it finishes, |numrpipes| records have been decoded. For each record 'i':
- * rr[i].type    - is the type of record
- * rr[i].data,   - data
- * rr[i].length, - number of bytes
- * Multiple records will only be returned if the record types are all
- * SSL3_RT_APPLICATION_DATA. The number of records returned will always be <=
- * |max_pipelines|
- */
-/* used only by ssl3_read_bytes */
-int ssl3_get_record(SSL_CONNECTION *s)
-{
-    int enc_err, rret;
-    int i;
-    size_t more, n;
-    SSL3_RECORD *rr, *thisrr;
-    SSL3_BUFFER *rbuf;
-    SSL_SESSION *sess;
-    unsigned char *p;
-    unsigned char md[EVP_MAX_MD_SIZE];
-    unsigned int version;
-    size_t mac_size = 0;
-    int imac_size;
-    size_t num_recs = 0, max_recs, j;
-    PACKET pkt, sslv2pkt;
-    int using_ktls;
-    SSL_MAC_BUF *macbufs = NULL;
-    int ret = -1;
-    SSL *ssl = SSL_CONNECTION_GET_SSL(s);
-
-    rr = RECORD_LAYER_get_rrec(&s->rlayer);
-    rbuf = s->rrlmethod->get0_rbuf(s->rrl);
-
-    max_recs = s->max_pipelines;
-    if (max_recs == 0)
-        max_recs = 1;
-    sess = s->session;
-
-    /*
-     * KTLS reads full records. If there is any data left,
-     * then it is from before enabling ktls.
-     */
-    using_ktls = BIO_get_ktls_recv(s->rbio) && SSL3_BUFFER_get_left(rbuf) == 0;
-
-    do {
-        thisrr = &rr[num_recs];
-
-        /* check if we have the header */
-        if ((RECORD_LAYER_get_rstate(&s->rlayer) != SSL_ST_READ_BODY) ||
-            (s->rrlmethod->get_packet_length(s->rrl) < SSL3_RT_HEADER_LENGTH)) {
-            size_t sslv2len;
-            unsigned int type;
-
-            rret = HANDLE_RLAYER_RETURN(s,
-                s->rrlmethod->read_n(s->rrl, SSL3_RT_HEADER_LENGTH,
-                                     SSL3_BUFFER_get_len(rbuf), 0,
-                                     num_recs == 0 ? 1 : 0, &n));
-
-            if (rret <= 0) {
-#ifndef OPENSSL_NO_KTLS
-                if (!using_ktls || rret == 0)
-                    return rret;     /* error or non-blocking */
-                switch (errno) {
-                case EBADMSG:
-                    SSLfatal(s, SSL_AD_BAD_RECORD_MAC,
-                             SSL_R_DECRYPTION_FAILED_OR_BAD_RECORD_MAC);
-                    break;
-                case EMSGSIZE:
-                    SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                             SSL_R_PACKET_LENGTH_TOO_LONG);
-                    break;
-                case EINVAL:
-                    SSLfatal(s, SSL_AD_PROTOCOL_VERSION,
-                             SSL_R_WRONG_VERSION_NUMBER);
-                    break;
-                default:
-                    break;
-                }
-#endif
-                return rret;
-            }
-            RECORD_LAYER_set_rstate(&s->rlayer, SSL_ST_READ_BODY);
-
-            p = s->rrlmethod->get0_packet(s->rrl);
-            if (!PACKET_buf_init(&pkt, p,
-                                 s->rrlmethod->get_packet_length(s->rrl))) {
-                SSLfatal(s, SSL_AD_INTERNAL_ERROR, ERR_R_INTERNAL_ERROR);
-                return -1;
-            }
-            sslv2pkt = pkt;
-            if (!PACKET_get_net_2_len(&sslv2pkt, &sslv2len)
-                    || !PACKET_get_1(&sslv2pkt, &type)) {
-                SSLfatal(s, SSL_AD_DECODE_ERROR, ERR_R_INTERNAL_ERROR);
-                return -1;
-            }
-            /*
-             * The first record received by the server may be a V2ClientHello.
-             */
-            if (s->server && RECORD_LAYER_is_first_record(&s->rlayer)
-                    && (sslv2len & 0x8000) != 0
-                    && (type == SSL2_MT_CLIENT_HELLO)) {
-                /*
-                 *  SSLv2 style record
-                 *
-                 * |num_recs| here will actually always be 0 because
-                 * |num_recs > 0| only ever occurs when we are processing
-                 * multiple app data records - which we know isn't the case here
-                 * because it is an SSLv2ClientHello. We keep it using
-                 * |num_recs| for the sake of consistency
-                 */
-                thisrr->type = SSL3_RT_HANDSHAKE;
-                thisrr->rec_version = SSL2_VERSION;
-
-                thisrr->length = sslv2len & 0x7fff;
-
-                if (thisrr->length > SSL3_BUFFER_get_len(rbuf)
-                    - SSL2_RT_HEADER_LENGTH) {
-                    SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                             SSL_R_PACKET_LENGTH_TOO_LONG);
-                    return -1;
-                }
-
-                if (thisrr->length < MIN_SSL2_RECORD_LEN) {
-                    SSLfatal(s, SSL_AD_DECODE_ERROR, SSL_R_LENGTH_TOO_SHORT);
-                    return -1;
-                }
-            } else {
-                /* SSLv3+ style record */
-
-                /* Pull apart the header into the SSL3_RECORD */
-                if (!PACKET_get_1(&pkt, &type)
-                        || !PACKET_get_net_2(&pkt, &version)
-                        || !PACKET_get_net_2_len(&pkt, &thisrr->length)) {
-                    if (s->msg_callback)
-                        s->msg_callback(0, 0, SSL3_RT_HEADER, p, 5, ssl,
-                                        s->msg_callback_arg);
-                    SSLfatal(s, SSL_AD_DECODE_ERROR, ERR_R_INTERNAL_ERROR);
-                    return -1;
-                }
-                thisrr->type = type;
-                thisrr->rec_version = version;
-
-                if (s->msg_callback)
-                    s->msg_callback(0, version, SSL3_RT_HEADER, p, 5, ssl,
-                                    s->msg_callback_arg);
-
-                /*
-                 * Lets check version. In TLSv1.3 we only check this field
-                 * when encryption is occurring (see later check). For the
-                 * ServerHello after an HRR we haven't actually selected TLSv1.3
-                 * yet, but we still treat it as TLSv1.3, so we must check for
-                 * that explicitly
-                 */
-                if (!s->first_packet && !SSL_CONNECTION_IS_TLS13(s)
-                        && s->hello_retry_request != SSL_HRR_PENDING
-                        && version != (unsigned int)s->version) {
-                    if ((s->version & 0xFF00) == (version & 0xFF00)
-                        && !s->enc_write_ctx && !s->write_hash) {
-                        if (thisrr->type == SSL3_RT_ALERT) {
-                            /*
-                             * The record is using an incorrect version number,
-                             * but what we've got appears to be an alert. We
-                             * haven't read the body yet to check whether its a
-                             * fatal or not - but chances are it is. We probably
-                             * shouldn't send a fatal alert back. We'll just
-                             * end.
-                             */
-                            SSLfatal(s, SSL_AD_NO_ALERT,
-                                     SSL_R_WRONG_VERSION_NUMBER);
-                            return -1;
-                        }
-                        /*
-                         * Send back error using their minor version number :-)
-                         */
-                        s->version = (unsigned short)version;
-                    }
-                    SSLfatal(s, SSL_AD_PROTOCOL_VERSION,
-                             SSL_R_WRONG_VERSION_NUMBER);
-                    return -1;
-                }
-
-                if ((version >> 8) != SSL3_VERSION_MAJOR) {
-                    if (RECORD_LAYER_is_first_record(&s->rlayer)) {
-                        /* Go back to start of packet, look at the five bytes
-                         * that we have. */
-                        p = s->rrlmethod->get0_packet(s->rrl);
-                        if (HAS_PREFIX((char *)p, "GET ") ||
-                            HAS_PREFIX((char *)p, "POST ") ||
-                            HAS_PREFIX((char *)p, "HEAD ") ||
-                            HAS_PREFIX((char *)p, "PUT ")) {
-                            SSLfatal(s, SSL_AD_NO_ALERT, SSL_R_HTTP_REQUEST);
-                            return -1;
-                        } else if (HAS_PREFIX((char *)p, "CONNE")) {
-                            SSLfatal(s, SSL_AD_NO_ALERT,
-                                     SSL_R_HTTPS_PROXY_REQUEST);
-                            return -1;
-                        }
-
-                        /* Doesn't look like TLS - don't send an alert */
-                        SSLfatal(s, SSL_AD_NO_ALERT,
-                                 SSL_R_WRONG_VERSION_NUMBER);
-                        return -1;
-                    } else {
-                        SSLfatal(s, SSL_AD_PROTOCOL_VERSION,
-                                 SSL_R_WRONG_VERSION_NUMBER);
-                        return -1;
-                    }
-                }
-
-                if (SSL_CONNECTION_IS_TLS13(s)
-                        && s->enc_read_ctx != NULL
-                        && !using_ktls) {
-                    if (thisrr->type != SSL3_RT_APPLICATION_DATA
-                            && (thisrr->type != SSL3_RT_CHANGE_CIPHER_SPEC
-                                || !SSL_IS_FIRST_HANDSHAKE(s))
-                            && (thisrr->type != SSL3_RT_ALERT
-                                || s->statem.enc_read_state
-                                   != ENC_READ_STATE_ALLOW_PLAIN_ALERTS)) {
-                        SSLfatal(s, SSL_AD_UNEXPECTED_MESSAGE,
-                                 SSL_R_BAD_RECORD_TYPE);
-                        return -1;
-                    }
-                    if (thisrr->rec_version != TLS1_2_VERSION) {
-                        SSLfatal(s, SSL_AD_DECODE_ERROR,
-                                 SSL_R_WRONG_VERSION_NUMBER);
-                        return -1;
-                    }
-                }
-
-                if (thisrr->length >
-                    SSL3_BUFFER_get_len(rbuf) - SSL3_RT_HEADER_LENGTH) {
-                    SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                             SSL_R_PACKET_LENGTH_TOO_LONG);
-                    return -1;
-                }
-            }
-
-            /* now s->rlayer.rstate == SSL_ST_READ_BODY */
-        }
-
-        if (SSL_CONNECTION_IS_TLS13(s)) {
-            size_t len = SSL3_RT_MAX_TLS13_ENCRYPTED_LENGTH;
-
-            /* KTLS strips the inner record type. */
-            if (using_ktls)
-                len = SSL3_RT_MAX_ENCRYPTED_LENGTH;
-
-            if (thisrr->length > len) {
-                SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                         SSL_R_ENCRYPTED_LENGTH_TOO_LONG);
-                return -1;
-            }
-        } else {
-            size_t len = SSL3_RT_MAX_ENCRYPTED_LENGTH;
-
-#ifndef OPENSSL_NO_COMP
-            /*
-             * If OPENSSL_NO_COMP is defined then SSL3_RT_MAX_ENCRYPTED_LENGTH
-             * does not include the compression overhead anyway.
-             */
-            if (s->expand == NULL)
-                len -= SSL3_RT_MAX_COMPRESSED_OVERHEAD;
-#endif
-
-            /* KTLS may use all of the buffer */
-            if (using_ktls)
-                len = SSL3_BUFFER_get_left(rbuf);
-
-            if (thisrr->length > len) {
-                SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                         SSL_R_ENCRYPTED_LENGTH_TOO_LONG);
-                return -1;
-            }
-        }
-
-        /*
-         * s->rlayer.rstate == SSL_ST_READ_BODY, get and decode the data.
-         * Calculate how much more data we need to read for the rest of the
-         * record
-         */
-        if (thisrr->rec_version == SSL2_VERSION) {
-            more = thisrr->length + SSL2_RT_HEADER_LENGTH
-                - SSL3_RT_HEADER_LENGTH;
-        } else {
-            more = thisrr->length;
-        }
-
-        if (more > 0) {
-            /* now s->rlayer.packet_length == SSL3_RT_HEADER_LENGTH */
-
-            rret = HANDLE_RLAYER_RETURN(s,
-                s->rrlmethod->read_n(s->rrl, more, more, 1, 0, &n));
-            if (rret <= 0)
-                return rret;     /* error or non-blocking io */
-        }
-
-        /* set state for later operations */
-        RECORD_LAYER_set_rstate(&s->rlayer, SSL_ST_READ_HEADER);
-
-        /*
-         * At this point, s->rlayer.packet_length == SSL3_RT_HEADER_LENGTH
-         * + thisrr->length, or s->rlayer.packet_length == SSL2_RT_HEADER_LENGTH
-         * + thisrr->length and we have that many bytes in s->rlayer.packet
-         */
-        if (thisrr->rec_version == SSL2_VERSION) {
-            thisrr->input =
-                &(s->rrlmethod->get0_packet(s->rrl)[SSL2_RT_HEADER_LENGTH]);
-        } else {
-            thisrr->input =
-                &(s->rrlmethod->get0_packet(s->rrl)[SSL3_RT_HEADER_LENGTH]);
-        }
-
-        /*
-         * ok, we can now read from 's->rlayer.packet' data into 'thisrr'.
-         * thisrr->input points at thisrr->length bytes, which need to be copied
-         * into thisrr->data by either the decryption or by the decompression.
-         * When the data is 'copied' into the thisrr->data buffer,
-         * thisrr->input will be updated to point at the new buffer
-         */
-
-        /*
-         * We now have - encrypted [ MAC [ compressed [ plain ] ] ]
-         * thisrr->length bytes of encrypted compressed stuff.
-         */
-
-        /* decrypt in place in 'thisrr->input' */
-        thisrr->data = thisrr->input;
-        thisrr->orig_len = thisrr->length;
-
-        /* Mark this record as not read by upper layers yet */
-        thisrr->read = 0;
-
-        num_recs++;
-
-        /* we have pulled in a full packet so zero things */
-        s->rrlmethod->reset_packet_length(s->rrl);
-        RECORD_LAYER_clear_first_record(&s->rlayer);
-    } while (num_recs < max_recs
-             && thisrr->type == SSL3_RT_APPLICATION_DATA
-             && SSL_USE_EXPLICIT_IV(s)
-             && s->enc_read_ctx != NULL
-             && (EVP_CIPHER_get_flags(EVP_CIPHER_CTX_get0_cipher(s->enc_read_ctx))
-                 & EVP_CIPH_FLAG_PIPELINE) != 0
-             && ssl3_record_app_data_waiting(s));
-
-    if (num_recs == 1
-            && thisrr->type == SSL3_RT_CHANGE_CIPHER_SPEC
-            && (SSL_CONNECTION_IS_TLS13(s)
-                || s->hello_retry_request != SSL_HRR_NONE)
-            && SSL_IS_FIRST_HANDSHAKE(s)) {
-        /*
-         * CCS messages must be exactly 1 byte long, containing the value 0x01
-         */
-        if (thisrr->length != 1 || thisrr->data[0] != 0x01) {
-            SSLfatal(s, SSL_AD_ILLEGAL_PARAMETER,
-                     SSL_R_INVALID_CCS_MESSAGE);
-            return -1;
-        }
-        /*
-         * CCS messages are ignored in TLSv1.3. We treat it like an empty
-         * handshake record
-         */
-        thisrr->type = SSL3_RT_HANDSHAKE;
-        RECORD_LAYER_inc_empty_record_count(&s->rlayer);
-        if (RECORD_LAYER_get_empty_record_count(&s->rlayer)
-            > MAX_EMPTY_RECORDS) {
-            SSLfatal(s, SSL_AD_UNEXPECTED_MESSAGE,
-                     SSL_R_UNEXPECTED_CCS_MESSAGE);
-            return -1;
-        }
-        thisrr->read = 1;
-        RECORD_LAYER_set_numrpipes(&s->rlayer, 1);
-
-        return 1;
-    }
-
-    if (using_ktls)
-        goto skip_decryption;
-
-    if (s->read_hash != NULL) {
-        const EVP_MD *tmpmd = EVP_MD_CTX_get0_md(s->read_hash);
-
-        if (tmpmd != NULL) {
-            imac_size = EVP_MD_get_size(tmpmd);
-            if (!ossl_assert(imac_size >= 0 && imac_size <= EVP_MAX_MD_SIZE)) {
-                    SSLfatal(s, SSL_AD_INTERNAL_ERROR, ERR_R_EVP_LIB);
-                    return -1;
-            }
-            mac_size = (size_t)imac_size;
-        }
-    }
-
-    /*
-     * If in encrypt-then-mac mode calculate mac from encrypted record. All
-     * the details below are public so no timing details can leak.
-     */
-    if (SSL_READ_ETM(s) && s->read_hash) {
-        unsigned char *mac;
-
-        for (j = 0; j < num_recs; j++) {
-            thisrr = &rr[j];
-
-            if (thisrr->length < mac_size) {
-                SSLfatal(s, SSL_AD_DECODE_ERROR, SSL_R_LENGTH_TOO_SHORT);
-                return -1;
-            }
-            thisrr->length -= mac_size;
-            mac = thisrr->data + thisrr->length;
-            i = ssl->method->ssl3_enc->mac(s, thisrr, md, 0 /* not send */ );
-            if (i == 0 || CRYPTO_memcmp(md, mac, mac_size) != 0) {
-                SSLfatal(s, SSL_AD_BAD_RECORD_MAC,
-                         SSL_R_DECRYPTION_FAILED_OR_BAD_RECORD_MAC);
-                return -1;
-            }
-        }
-        /*
-         * We've handled the mac now - there is no MAC inside the encrypted
-         * record
-         */
-        mac_size = 0;
-    }
-
-    if (mac_size > 0) {
-        macbufs = OPENSSL_zalloc(sizeof(*macbufs) * num_recs);
-        if (macbufs == NULL) {
-            SSLfatal(s, SSL_AD_INTERNAL_ERROR, ERR_R_MALLOC_FAILURE);
-            return -1;
-        }
-    }
-
-    enc_err = ssl->method->ssl3_enc->enc(s, rr, num_recs, 0, macbufs, mac_size);
-
-    /*-
-     * enc_err is:
-     *    0: if the record is publicly invalid, or an internal error, or AEAD
-     *       decryption failed, or ETM decryption failed.
-     *    1: Success or MTE decryption failed (MAC will be randomised)
-     */
-    if (enc_err == 0) {
-        if (ossl_statem_in_error(s)) {
-            /* SSLfatal() already got called */
-            goto end;
-        }
-        if (num_recs == 1 && ossl_statem_skip_early_data(s)) {
-            /*
-             * Valid early_data that we cannot decrypt will fail here. We treat
-             * it like an empty record.
-             */
-
-            thisrr = &rr[0];
-
-            if (!ossl_early_data_count_ok(s, thisrr->length,
-                                          EARLY_DATA_CIPHERTEXT_OVERHEAD, 0)) {
-                /* SSLfatal() already called */
-                goto end;
-            }
-
-            thisrr->length = 0;
-            thisrr->read = 1;
-            RECORD_LAYER_set_numrpipes(&s->rlayer, 1);
-            RECORD_LAYER_reset_read_sequence(&s->rlayer);
-            ret = 1;
-            goto end;
-        }
-        SSLfatal(s, SSL_AD_BAD_RECORD_MAC,
-                 SSL_R_DECRYPTION_FAILED_OR_BAD_RECORD_MAC);
-        goto end;
-    }
-    OSSL_TRACE_BEGIN(TLS) {
-        BIO_printf(trc_out, "dec %lu\n", (unsigned long)rr[0].length);
-        BIO_dump_indent(trc_out, rr[0].data, rr[0].length, 4);
-    } OSSL_TRACE_END(TLS);
-
-    /* r->length is now the compressed data plus mac */
-    if ((sess != NULL)
-            && (s->enc_read_ctx != NULL)
-            && (!SSL_READ_ETM(s) && EVP_MD_CTX_get0_md(s->read_hash) != NULL)) {
-        /* s->read_hash != NULL => mac_size != -1 */
-
-        for (j = 0; j < num_recs; j++) {
-            SSL_MAC_BUF *thismb = &macbufs[j];
-            thisrr = &rr[j];
-
-            i = ssl->method->ssl3_enc->mac(s, thisrr, md, 0 /* not send */ );
-            if (i == 0 || thismb == NULL || thismb->mac == NULL
-                || CRYPTO_memcmp(md, thismb->mac, (size_t)mac_size) != 0)
-                enc_err = 0;
-            if (thisrr->length > SSL3_RT_MAX_COMPRESSED_LENGTH + mac_size)
-                enc_err = 0;
-        }
-    }
-
-    if (enc_err == 0) {
-        if (ossl_statem_in_error(s)) {
-            /* We already called SSLfatal() */
-            goto end;
-        }
-        /*
-         * A separate 'decryption_failed' alert was introduced with TLS 1.0,
-         * SSL 3.0 only has 'bad_record_mac'.  But unless a decryption
-         * failure is directly visible from the ciphertext anyway, we should
-         * not reveal which kind of error occurred -- this might become
-         * visible to an attacker (e.g. via a logfile)
-         */
-        SSLfatal(s, SSL_AD_BAD_RECORD_MAC,
-                 SSL_R_DECRYPTION_FAILED_OR_BAD_RECORD_MAC);
-        goto end;
-    }
-
- skip_decryption:
-
-    for (j = 0; j < num_recs; j++) {
-        thisrr = &rr[j];
-
-        /* thisrr->length is now just compressed */
-        if (s->expand != NULL) {
-            if (thisrr->length > SSL3_RT_MAX_COMPRESSED_LENGTH) {
-                SSLfatal(s, SSL_AD_RECORD_OVERFLOW,
-                         SSL_R_COMPRESSED_LENGTH_TOO_LONG);
-                goto end;
-            }
-            if (!ssl3_do_uncompress(s, thisrr)) {
-                SSLfatal(s, SSL_AD_DECOMPRESSION_FAILURE,
-                         SSL_R_BAD_DECOMPRESSION);
-                goto end;
-            }
-        }
-
-        if (SSL_CONNECTION_IS_TLS13(s)
-                && s->enc_read_ctx != NULL
-                && thisrr->type != SSL3_RT_ALERT) {
-            /*
-             * The following logic are irrelevant in KTLS: the kernel provides
-             * unprotected record and thus record type represent the actual
-             * content type, and padding is already removed and thisrr->type and
-             * thisrr->length should have the correct values.
-             */
-            if (!using_ktls) {
-                size_t end;
-
-                if (thisrr->length == 0
-                        || thisrr->type != SSL3_RT_APPLICATION_DATA) {
-                    SSLfatal(s, SSL_AD_UNEXPECTED_MESSAGE, SSL_R_BAD_RECORD_TYPE);
-                    goto end;
-                }
-
-                /* Strip trailing padding */
-                for (end = thisrr->length - 1; end > 0 && thisrr->data[end] == 0;
-                     end--)
-                    continue;
-
-                thisrr->length = end;
-                thisrr->type = thisrr->data[end];
-            }
-            if (thisrr->type != SSL3_RT_APPLICATION_DATA
-                    && thisrr->type != SSL3_RT_ALERT
-                    && thisrr->type != SSL3_RT_HANDSHAKE) {
-                SSLfatal(s, SSL_AD_UNEXPECTED_MESSAGE, SSL_R_BAD_RECORD_TYPE);
-                goto end;
-            }
-            if (s->msg_callback)
-                s->msg_callback(0, s->version, SSL3_RT_INNER_CONTENT_TYPE,
-                                &thisrr->type, 1, ssl, s->msg_callback_arg);
-        }
-
-        /*
-         * TLSv1.3 alert and handshake records are required to be non-zero in
-         * length.
-         */
-        if (SSL_CONNECTION_IS_TLS13(s)
-                && (thisrr->type == SSL3_RT_HANDSHAKE
-                    || thisrr->type == SSL3_RT_ALERT)
-                && thisrr->length == 0) {
-            SSLfatal(s, SSL_AD_UNEXPECTED_MESSAGE, SSL_R_BAD_LENGTH);
-            goto end;
-        }
-
-        /*
-         * Usually thisrr->length is the length of a single record, but when
-         * KTLS handles the decryption, thisrr->length may be larger than
-         * SSL3_RT_MAX_PLAIN_LENGTH because the kernel may have coalesced
-         * multiple records.
-         * Therefore we have to rely on KTLS to check the plaintext length
-         * limit in the kernel.
-         */
-        if (thisrr->length > SSL3_RT_MAX_PLAIN_LENGTH && !using_ktls) {
-            SSLfatal(s, SSL_AD_RECORD_OVERFLOW, SSL_R_DATA_LENGTH_TOO_LONG);
-            goto end;
-        }
-
-        /*
-         * Check if the received packet overflows the current
-         * Max Fragment Length setting.
-         * Note: USE_MAX_FRAGMENT_LENGTH_EXT and KTLS are mutually exclusive.
-         */
-        if (s->session != NULL && USE_MAX_FRAGMENT_LENGTH_EXT(s->session)
-                && thisrr->length > GET_MAX_FRAGMENT_LENGTH(s->session)) {
-            SSLfatal(s, SSL_AD_RECORD_OVERFLOW, SSL_R_DATA_LENGTH_TOO_LONG);
-            goto end;
-        }
-
-        thisrr->off = 0;
-        /*-
-         * So at this point the following is true
-         * thisrr->type   is the type of record
-         * thisrr->length == number of bytes in record
-         * thisrr->off    == offset to first valid byte
-         * thisrr->data   == where to take bytes from, increment after use :-).
-         */
-
-        /* just read a 0 length packet */
-        if (thisrr->length == 0) {
-            RECORD_LAYER_inc_empty_record_count(&s->rlayer);
-            if (RECORD_LAYER_get_empty_record_count(&s->rlayer)
-                > MAX_EMPTY_RECORDS) {
-                SSLfatal(s, SSL_AD_UNEXPECTED_MESSAGE, SSL_R_RECORD_TOO_SMALL);
-                goto end;
-            }
-        } else {
-            RECORD_LAYER_reset_empty_record_count(&s->rlayer);
-        }
-    }
-
-    if (s->early_data_state == SSL_EARLY_DATA_READING) {
-        thisrr = &rr[0];
-        if (thisrr->type == SSL3_RT_APPLICATION_DATA
-                && !ossl_early_data_count_ok(s, thisrr->length, 0, 0)) {
-            /* SSLfatal already called */
-            goto end;
-        }
-    }
-
-    RECORD_LAYER_set_numrpipes(&s->rlayer, num_recs);
-    ret = 1;
- end:
-    if (macbufs != NULL) {
-        for (j = 0; j < num_recs; j++) {
-            if (macbufs[j].alloced)
-                OPENSSL_free(macbufs[j].mac);
-        }
-        OPENSSL_free(macbufs);
-    }
-    return ret;
-}
 
 int ssl3_do_uncompress(SSL_CONNECTION *sc, SSL3_RECORD *rr)
 {
