+++ test/sslapitest.c
@@ -1541,7 +1541,6 @@ static int execute_cleanse_plaintext(const SSL_METHOD *smeth,
     SSL_CTX *cctx = NULL, *sctx = NULL;
     SSL *clientssl = NULL, *serverssl = NULL;
     int testresult = 0;
-    SSL3_RECORD *rr;
     void *zbuf;
     SSL_CONNECTION *serversc;
 
@@ -1606,10 +1605,24 @@ static int execute_cleanse_plaintext(const SSL_METHOD *smeth,
      */
     if (!TEST_ptr(serversc = SSL_CONNECTION_FROM_SSL_ONLY(serverssl)))
         goto end;
-    rr = serversc->rlayer.rrec;
-    zbuf = &rr->data[rr->off];
-    if (!TEST_int_eq(rr->length, sizeof(cbuf)))
-        goto end;
+
+    /*
+     * TODO(RECLAYER): This is temporary until DTLS is converted to use the new
+     * record layer code.
+     */
+    if (!SSL_is_dtls(serverssl)) {
+        TLS_RECORD *rr = serversc->rlayer.tlsrecs;
+
+        zbuf = &rr->data[rr->off];
+        if (!TEST_int_eq(rr->length, sizeof(cbuf)))
+            goto end;
+    } else {
+        SSL3_RECORD *rr = serversc->rlayer.rrec;
+
+        zbuf = &rr->data[rr->off];
+        if (!TEST_int_eq(rr->length, sizeof(cbuf)))
+            goto end;
+    }
 
     /*
      * After SSL_peek() the plaintext must still be stored in the
